+++ src/test/java/kr/hhplus/be/server/config/redis/EmbeddedRedisConfig.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.config;
+package kr.hhplus.be.server.config.redis;
 
 import jakarta.annotation.PostConstruct;
 import jakarta.annotation.PreDestroy;
