+++ src/test/java/kr/hhplus/be/server/config/redis/TestRedisConfig.java
@@ -1,10 +1,9 @@
-package kr.hhplus.be.server.config;
+package kr.hhplus.be.server.config.redis;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.redis.connection.RedisConnectionFactory;
 import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.data.redis.serializer.GenericJackson2JsonRedisSerializer;
 import org.springframework.data.redis.serializer.StringRedisSerializer;
 
 @Configuration
