+++ src/test/java/kr/hhplus/be/server/config/redis/TestRedissonConfig.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.config;
+package kr.hhplus.be.server.config.redis;
 
 import org.redisson.Redisson;
 import org.redisson.api.RedissonClient;
